requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Database & ORM
sqlalchemy
alembic
psycopg2-binary

# AI
google-generativeai>=0.3.0
langchain
langchain-google-genai>=0.0.8
langchain-community
duckduckgo-search

# Utilities
python-dotenv
pydantic
pydantic-settings
requests

# Web Scraping
playwright
beautifulsoup4
